cmd: split artifactgen into its own binary
The `artifacts` subcommand of `tokengen` is the only subcommand that
pulls the Fabric Smart Client `integration/nwo` topology framework (and
its transitive Google Cloud / bench tooling dependencies) into the
`tokengen` binary. The other subcommands (`pp`, `certfier`, `version`)
do not need any of that.

Move the artifact generator into a standalone `cmd/artifactgen` binary
and drop it from `tokengen`'s root command, as suggested by @adecaro
on #1445. No functional changes to the subcommand itself -- the files
are moved with `git mv` so history is preserved.

Effect on `cmd/tokengen`'s transitive dependency graph:
  1762 -> 869 packages (~51% reduction).

`cmd/tokengen` still pulls a subset of `integration/nwo` via `pp`
(chaincode packager) and `certfier` (topology-aware keygen). Slimming
those is a separate concern and is intentionally out of scope here.

Changes:
  * move cmd/tokengen/cobra/artifactgen/** to cmd/artifactgen/**
  * add cmd/artifactgen/main.go wiring the `artifacts` subcommand
  * repackage root-level moved files as `package main`
  * drop artifactgen wiring from cmd/tokengen/main.go
  * add `make artifactgen` target alongside `make tokengen`
  * update READMEs on both sides, with a cross-link from tokengen

Fixes #1445
