+++ cmd/artifactgen/testdata/topology.go
